+++ src/validators/boolean/boolean.validator.test.js
@@ -1,20 +1,6 @@
 const { createContext } = require('../../context')
 const validator = require('./boolean.validator')
 
-test('allow null value should return no errors', () => {
-  const ctx = createContext({
-    schema: {
-      $type: 'boolean',
-      allow: [null]
-    },
-    value: null
-  })
-
-  validator.validate(ctx)
-
-  expect(ctx.valid).toBeTruthy()
-})
-
 test('invalid type should return errors', () => {
   const ctx = createContext({
     schema: {