+++ src/validators/custom/custom.validator.js
@@ -1,8 +1,6 @@
 const customValidator = {
   $type: 'custom',
   validate: ctx => {
-    if (ctx.handlePreValidation().handled) return
-
     const validationResult = ctx.schema.validation(ctx.value)
 
     if (!validationResult.valid) {