+++ tasks/generate-docs/generate-docs.js
@@ -1,3 +1,5 @@
 const generateValidators = require('./validators')
+const generateMiddlewares = require('./middlewares')
 
 generateValidators()
+generateMiddlewares()